coalescewalletdata
