+++ packages/app-mobile/package.json
@@ -11,8 +11,8 @@
     "gql:generate": "graphql-codegen --config codegen.yml",
     "ios": "APP_ENV=development EXPO_USE_CUSTOM_INSPECTOR_PROXY=true expo start --ios --dev-client",
     "ios:deploy": "eas build --profile production --platform ios --auto-submit --non-interactive",
-    "ios:dev-client": "APP_ENV=development eas build --profile development-simulator --platform ios",
-    "ios:dev-phone": "APP_ENV=development eas build --profile development --platform ios",
+    "ios:dev-client": "eas build --profile development-simulator --platform ios",
+    "ios:dev-phone": "eas build --profile development --platform ios",
     "web": "expo start --web",
     "add": "expo add",
     "eject": "expo eject",
